#167 - Implemented suggested changes
